requirements_jetson.txt
absl-py==0.11.0
aiohttp==3.7.4
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
charset-normalizer==2.0.12
click==7.1.2
cycler==0.10.0
einops==0.3.0
filelock==3.0.12
fonttools==4.24.0
frozenlist==1.2.0
fsspec==2021.4.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
grpcio-status==1.34.0
h5py==2.10.0
huggingface-hub==0.0.8
hydra-core==1.0.7
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.10.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
kornia==0.5.0
lmdb==1.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
# matplotlib==2.1.1
# sudo chown -R $USER:$USER /home/stormsight/DP/StormSight/torch_env_new
# pip install matplotlib==3.1.3 --no-cache-dir
# ml-dtypes==0.0.2
mpmath==1.2.1
multidict==5.1.0
natsort==7.1.1
networkx==2.5.1
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.6
# opencv-python -> To install manually with CUDA later
packaging==20.9
Pillow==8.4.0
protobuf==3.17.3
psutil==5.8.0
ptflops==0.6.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.3.8
torchmetrics==0.6.2
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
safetensors==0.2.7
scikit-image==0.17.2
scipy==1.5.4
setuptools==49.6.0
shapely==1.8.5
six==1.15.0
sympy==1.6.2
tensorboard==2.4.1
tensorboard-data-server==0.6.0
tensorboardx==2.1
tifffile==2020.9.3
timm==0.3.2
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
warmup-scheduler==0.3
Werkzeug==1.0.1
wheel==0.36.2
yacs==0.1.8
yarl==1.6.3
zipp==3.4.1
nanocamera

# torch torchvision torchaudio opencv (all with cuda)
